# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AppCloud API",
    "description": "Knowledge graph platform for infrastructure dependency mapping and blast-radius analysis.",
    "version": "1.1.0"
  },
  "components": {
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
...
  },
  "paths": {
...
    "/integrations/cloud": {
      "get": {
        "summary": "List all cloud accounts (decrypted config)",
        "tags": [
          "Integrations"
        ],
        "description": "Returns every cloud account row with its decrypted `config` (the on-disk row stores AES-256-GCM-encrypted secrets — clientSecret, secretAccessKey, private_key — which are decrypted on read).",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true,
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "provider": {
                        "type": "string",
                        "enum": [
                          "aws",
                          "azure",
                          "gcp"
                        ]
                      },
                      "name": {
                        "type": "string"
                      },
                      "config": {
                        "type": "object",
                        "additionalProperties": true,
                        "description": "Provider-specific credentials. Secret fields (clientSecret, secretAccessKey, private_key) are AES-256-GCM encrypted at rest."
                      },
                      "enabled": {
                        "type": "boolean"
                      },
                      "last_scan_at": {
                        "type": [
                          "null",
                          "string"
                        ],
                        "format": "date-time"
                      },
                      "last_scan_status": {
                        "type": [
                          "null",
                          "string"
                        ]
                      },
                      "last_scan_total": {
                        "type": [
                          "null",
                          "integer"
                        ]
                      },
                      "last_scan_error": {
                        "type": [
                          "null",
                          "string"
                        ]
                      },
                      "created_at": {
                        "type": [
                          "null",
                          "string"
                        ],
                        "format": "date-time"
                      },
                      "updated_at": {
                        "type": [
                          "null",
                          "string"
                        ],
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create or upsert a cloud account",
        "tags": [
          "Integrations"
        ],
        "description": "On unique-key conflict (provider, name) the row is updated, not duplicated. Secret fields in `config` (clientSecret, secretAccessKey, private_key) are AES-256-GCM-encrypted before INSERT.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "provider",
                  "name"
                ],
                "additionalProperties": true,
                "properties": {
                  "provider": {
                    "type": "string",
                    "enum": [
                      "aws",
                      "azure",
                      "gcp"
                    ]
                  },
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "config": {
                    "type": "object",
                    "additionalProperties": true,
                    "description": "Azure: { subscriptionId, tenantId, clientId, clientSecret }. GCP: { projectId, serviceAccount: <JSON string> }. AWS: { accessKeyId, secretAccessKey, regions, aggregatorName, aggregatorRegion }."
                  },
                  "enabled": {
                    "type": "boolean"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "provider": {
                      "type": "string",
                      "enum": [
                        "aws",
                        "azure",
                        "gcp"
                      ]
                    },
                    "name": {
                      "type": "string"
                    },
                    "config": {
                      "type": "object",
                      "additionalProperties": true,
                      "description": "Provider-specific credentials. Secret fields (clientSecret, secretAccessKey, private_key) are AES-256-GCM encrypted at rest."
                    },
                    "enabled": {
                      "type": "boolean"
                    },
                    "last_scan_at": {
                      "type": [
                        "null",
                        "string"
                      ],
                      "format": "date-time"
                    },
                    "last_scan_status": {
                      "type": [
                        "null",
                        "string"
                      ]
                    },
                    "last_scan_total": {
                      "type": [
                        "null",
                        "integer"
                      ]
                    },
                    "last_scan_error": {
                      "type": [
                        "null",
                        "string"
                      ]
                    },
                    "created_at": {
                      "type": [
                        "null",
                        "string"
                      ],
                      "format": "date-time"
                    },
                    "updated_at": {
                      "type": [
                        "null",
                        "string"
                      ],
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "error",
                    "message"
                  ],
                  "additionalProperties": true,
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 400
                    },
                    "error": {
                      "type": "string",
                      "example": "Bad Request"
                    },
                    "message": {
                      "type": "string",
                      "example": "infraId required"
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "error",
                    "message"
                  ],
                  "additionalProperties": true,
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 400
                    },
                    "error": {
                      "type": "string",
                      "example": "Bad Request"
                    },
                    "message": {
                      "type": "string",
                      "example": "infraId required"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/integrations/cloud/{provider}": {
      "get": {
        "summary": "List enabled cloud accounts for a provider",
        "tags": [
          "Integrations"
        ],
        "description": "Filters by provider (aws / azure / gcp) and `enabled = true`. Discovery scans use this same query indirectly via the `loadAccounts` helper.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "aws",
                "azure",
                "gcp"
              ]
            },
            "in": "path",
            "name": "provider",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true,
                    "properties": {
                      "id": {
                        "type": "string",
                        "format": "uuid"
                      },
                      "provider": {
                        "type": "string",
                        "enum": [
                          "aws",
                          "azure",
                          "gcp"
                        ]
                      },
                      "name": {
                        "type": "string"
                      },
                      "config": {
                        "type": "object",
                        "additionalProperties": true,
                        "description": "Provider-specific credentials. Secret fields (clientSecret, secretAccessKey, private_key) are AES-256-GCM encrypted at rest."
                      },
                      "enabled": {
                        "type": "boolean"
                      },
                      "last_scan_at": {
                        "type": [
                          "null",
                          "string"
                        ],
                        "format": "date-time"
                      },
                      "last_scan_status": {
                        "type": [
                          "null",
                          "string"
                        ]
                      },
                      "last_scan_total": {
                        "type": [
                          "null",
                          "integer"
                        ]
                      },
                      "last_scan_error": {
                        "type": [
                          "null",
                          "string"
                        ]
                      },
                      "created_at": {
                        "type": [
                          "null",
                          "string"
                        ],
                        "format": "date-time"
                      },
                      "updated_at": {
                        "type": [
                          "null",
                          "string"
                        ],
                        "format": "date-time"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/integrations/cloud/{id}": {
      "patch": {
        "summary": "Update a cloud account (partial)",
        "tags": [
          "Integrations"
        ],
        "description": "Any subset of `name` / `config` / `enabled`. The `config` shape is *merged* into the existing decrypted config and re-encrypted, so you can patch one secret without re-supplying the others.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "config": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "enabled": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Resource UUID or unique name"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "provider": {
                      "type": "string",
                      "enum": [
                        "aws",
                        "azure",
                        "gcp"
                      ]
                    },
                    "name": {
                      "type": "string"
                    },
                    "config": {
                      "type": "object",
                      "additionalProperties": true,
                      "description": "Provider-specific credentials. Secret fields (clientSecret, secretAccessKey, private_key) are AES-256-GCM encrypted at rest."
                    },
                    "enabled": {
                      "type": "boolean"
                    },
                    "last_scan_at": {
                      "type": [
                        "null",
                        "string"
                      ],
                      "format": "date-time"
                    },
                    "last_scan_status": {
                      "type": [
                        "null",
                        "string"
                      ]
                    },
                    "last_scan_total": {
                      "type": [
                        "null",
                        "integer"
                      ]
                    },
                    "last_scan_error": {
                      "type": [
                        "null",
                        "string"
                      ]
                    },
                    "created_at": {
                      "type": [
                        "null",
                        "string"
                      ],
                      "format": "date-time"
                    },
                    "updated_at": {
                      "type": [
                        "null",
                        "string"
                      ],
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "error",
                    "message"
                  ],
                  "additionalProperties": true,
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 400
                    },
                    "error": {
                      "type": "string",
                      "example": "Bad Request"
                    },
                    "message": {
                      "type": "string",
                      "example": "infraId required"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a cloud account",
        "tags": [
          "Integrations"
        ],
        "description": "Removes the Postgres row. Discovery-emitted Infra nodes already in Neo4j are *not* deleted; they'll be marked stale on the next scan that runs without this account.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Resource UUID or unique name"
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response"
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "error",
                    "message"
                  ],
                  "additionalProperties": true,
                  "properties": {
                    "statusCode": {
                      "type": "integer",
                      "example": 400
                    },
                    "error": {
                      "type": "string",
                      "example": "Bad Request"
                    },
                    "message": {
                      "type": "string",
                      "example": "infraId required"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/integrations/cloud/{id}/scan-result": {
      "patch": {
        "summary": "Internal — record a scan outcome on a cloud account",
        "tags": [
          "Integrations"
        ],
        "description": "Called by the scan handlers in discovery.js to stamp `last_scan_at / last_scan_status / last_scan_total / last_scan_error` on the account row. Not typically called directly by external consumers.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "success",
                      "error"
                    ]
                  },
                  "total": {
                    "type": "integer"
                  },
                  "error": {
                    "type": [
                      "string",
                      "null"
                    ]
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Resource UUID or unique name"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "updated": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    },
    {
      "url": "http://appcloud.local"
    }
  ],
  "security": [
    {
      "ApiKey": []
    }
  ]
}